requirements.txt
streamlit
streamlit-option-menu
io
Pillow
requests